+++ src/Modules/Mailbox/Mailbox.js
@@ -11,8 +11,6 @@ import { expose } from '@/utils'
 import {
   ajreq,
   hideLoader,
-  showLoader,
-  u_loadCoords,
   pulseInfo,
   pulseError,
   checkEmail
@@ -20,7 +18,6 @@ import {
 import './Mailbox.css'
 
 expose({
-  u_getGeo,
   mb_finishFile,
   mb_removeLast,
   mb_new_message,
@@ -33,30 +30,12 @@ expose({
   mb_clearEditor,
   mb_closeEditor,
   mb_send_message,
-  u_goAll,
   mb_refresh,
   checkEmail,
   u_handleNewEmail,
   u_addTypeHead,
   setAutocompleteAddresses
 })
-
-function u_getGeo (id) {
-  showLoader()
-
-  if ($(`#fs${id}plz`).val() != '' && $('#fs' + id + 'stadt').val() != '' && $('#fs' + id + 'anschrift').val() != '') {
-    u_loadCoords({
-      plz: $(`#fs${id}plz`).val(),
-      stadt: $(`#fs${id}stadt`).val(),
-      anschrift: $(`#fs${id}anschrift`).val(),
-      complete: function () {
-        hideLoader()
-      }
-    }, function (lat, lon) {
-      ajreq('updateGeo', { lat: lat, lon: lon, id: id })
-    })
-  }
-}
 
 function mb_finishFile (newname) {
   $('ul#et-file-list li:last').addClass('finish').append(`<input type="hidden" class="tmp" value="${newname}" name="tmp_${$('ul#et-file-list li').length}" />`)
@@ -202,10 +181,6 @@ function mb_send_message () {
       reply: parseInt($('#edit-reply').val())
     }, 'post')
   }
-}
-
-function u_goAll () {
-
 }
 
 function mb_refresh () {