+++ CHANGELOG.md
@@ -93,6 +93,7 @@ We are very proud to finally release foodsharing with an **AGPLv3** licence, mak
 - user normalisation in conversations API endpoint
 - remove unused quickprofile method !755 @NerdyProjects
 - fix a few linter warnings !755 @NerdyProjects
+- removed the geoClean and LostRegion modules !756 #103 @peter.toennies
 
 ## Dev/Test/CI stuff
 - better webpack splitting !681 @nicksellen